tools/xenstore: Add -Wno-bool-operation
This suppresses a warning from GCC7
  xenstored_domain.c:949:32: error: increment of a boolean expression

We have not yet 100% concluded that this should not be fixed by
changing the `_Bool b; b++;' to `b = 1'.  But this fixes the warning
and thus the build for now, and we want to apply the XSA-206 patches
to stable trees.

In the stable trees, this should be applied _before_ the XSA-206
series.

Reported-by: Michael Young <m.a.young@durham.ac.uk>
CC: Wei Liu <wei.liu2@citrix.com>
CC: Jan Beulich <JBeulich@suse.com>
Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>