ARM: vGIC: rework gic_raise_*_irq() functions
Currently gic_raise_inflight_irq() and gic_raise_guest_irq() are used
to queue interrupts to a VCPU, for which they are given a VCPU and an
interrupt number.
To allow proper locking and simplify further changes, change their prototypes
to take a struct pending_irq directly (since the callers have the
pointer already anyway).

Signed-off-by: Andre Przywara <andre.przywara@arm.com>