ARM: vGIC: move target vcpu from irq_rank to struct pending_irq
So far we kept the target VCPU for SPIs in the rank structure.
Move that information over into pending_irq.
This changes vgic_get_target_vcpu(), which now takes only a domain
and a struct pending_irq to get the target vCPU, in a way that does not
necessarily require the pending_irq lock to be held.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>