x86/string: Clean up the declarations of __builtin_mem*()
None of the GCC docs mention memmove() in its list of builtins even today, but
4.1 does have the builtin, meaning that all currently supported compilers have
it.

Rather than #undef'ing macros to avoid altering the function names, use the
method recommended by the C specification by enclosing the function name in
brackets to avoid the macro being expanded.  This means that optimisation
opportunities continue to work in the rest of the translation unit.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>