tools/firmware: pass EXTRAVERSION to seabios build
Having a fixed version string in the seabios binary is cumbersome:
 - it is required to write something to '.version' in the toplevel
   directory of seabios-dir-remove.
 - it is required to pass EXTRAVERSION= to make.

The latter is impossible without this change. For some reason an 'export
EXTRAVERSION=something' prior to 'make tools' is not sufficient. It has
to be passed in as cmdline option to make.

Add a make variable SEABIOS_EXTRAVERSION= and pass it to make.
Allow changing the default via the environment.
Document the new variable in INSTALL.

Signed-off-by: Olaf Hering <olaf@aepfle.de>