VT-d: fix VF of RC integrated endpoint matched to wrong VT-d unit
The problem is a VF of RC integrated PF (e.g. PF's BDF is 00:02.0),
we would wrongly use 00:00.0 to search VT-d unit.

To search VT-d unit for a VF, the BDF of the PF is used. And If the
PF is an Extended Function, the BDF of one traditional function is
used. The following line (from acpi_find_matched_drhd_unit()):
    devfn = PCI_SLOT(pdev->info.physfn.devfn) ? 0 : pdev->info.physfn.devfn;
sets 'devfn' to 0 if PF's devfn > 8. Apparently, this line treats all
PFs as ARI-capable function and assumes the Root Port or Switch
Downstream Port immediately above the PF has ARIforwarding enabled.
However, according to SRIOV spec 3.7.3, ARI is not applicable to RC
integrated PF. For this case, we should use PF's BDF directly other
than using 0 as devfn.

This patch adds a new pdev type to indicate a function is RC
integrated. And check whether PF is a RC integrated endpoint when
searching VT-d unit.

Reported-by: Crawford, Eric R <Eric.R.Crawford@intel.com>
Signed-off-by: Chao Gao <chao.gao@intel.com>