libxl: upstream Qemu as PV backend for IOEMU stubdom
Set the PV model version for the IOEMU stubdom from
the new device_model_pv_version setting. Existing code
sets up the command line and spawns the specified version
correctly.

In addition, the upstream QEMU doesn't allow the restore
console to be undefined (it SEGV's) so if this is not a restore
set it to "null" instead. This has the same affect on both upstream
QEMU and qemu-xen-traditional.

Signed-off-by: Simon Waterman <watermansrdev@gmail.com>